backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pytesseract
Pillow
PyMuPDF
tqdm
chardet
chromadb
openai
streamlit
